[Modify] SelectWordNumViewController
